+++ logic/models/task.go
@@ -1,6 +1,6 @@
 package models
 
 type Task struct {
-	Id   int32  `json:"id"`
+	Id   uint64	`json:"id"`
 	Name string `json:"name"`
 }
